// core/fetchers/fetchIgaListings.types.ts
export interface IgaResult {
  ProductId: string
  BrandName: string
  FullDisplayName: string
  IsAgeRequired: boolean
  SizeLabel: string
  Size: string
  ProductUrl: string
  ProductImageUrl: string
  HasNewPrice: boolean
  PromotionName?: any
  RegularPrice: number
  SalesPrice?: any
  CustomerProductComment: string
  PopularityFactor: number
}